attempt to fix multiple compile&enforcer-problems
