+++ misc/de-metas-common/de-metas-common-rest_api/pom.xml
@@ -38,7 +38,6 @@
 
     <properties>
         <metasfresh.version>10.0.0</metasfresh.version>
-        <jackson.version>2.10.4</jackson.version>
     </properties>
 
     <dependencies>
