+++ misc/de-metas-common/de-metas-common-shipmentschedule/src/main/java/de/metas/common/shipmentschedule/JsonRequestShipmentScheduleResultList.java
@@ -20,7 +20,7 @@
  * #L%
  */
 
-package de.metas.common.shipmentschedule.shipmentschedule;
+package de.metas.common.shipmentschedule;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -32,13 +32,14 @@ import lombok.Value;
 import java.util.List;
 
 @Value
-public class JsonResponseShipmentScheduleList
+public class JsonRequestShipmentScheduleResultList
 {
-	List<JsonResponseShipmentSchedule> items;
+	List<JsonRequestShipmentScheduleResult> items;
 
-	@Builder
 	@JsonCreator
-	private JsonResponseShipmentScheduleList(@JsonProperty("items") @Singular @NonNull final List<JsonResponseShipmentSchedule> items)
+	@Builder
+	private JsonRequestShipmentScheduleResultList(
+			@JsonProperty("items") @NonNull @Singular final List<JsonRequestShipmentScheduleResult> items)
 	{
 		this.items = items;
 	}
