+++ misc/de-metas-common/de-metas-common-shipmentschedule/src/main/java/de/metas/common/shipmentschedule/JsonResponseShipmentSchedule.java
@@ -20,7 +20,7 @@
  * #L%
  */
 
-package de.metas.common.shipmentschedule.shipmentschedule;
+package de.metas.common.shipmentschedule;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
