+++ misc/services/camel/de-metas-camel-shipmentschedule/src/main/java/de/metas/camel/shipment/JsonToXmlRouteBuilder.java
@@ -26,7 +26,7 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import de.metas.commons.shipmentschedule.JsonResponseShipmentScheduleList;
+import de.metas.common.shipmentschedule.JsonResponseShipmentScheduleList;
 import org.apache.camel.Exchange;
 import org.apache.camel.LoggingLevel;
 import org.apache.camel.builder.endpoint.EndpointRouteBuilder;
